+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## Unreleased
+
+### Features
+
+* Added ability to define a custom `logEntryDataTransformer` that allows you to transform the log entry metadata 
+  and data object before they are converted to canonical Entry object.
+
 ### [2.0.1](https://github.com/metcoder95/cloud-pine/compare/v2.0.0...v2.0.1) (2024-01-26)
 
 ## [2.0.0](https://github.com/metcoder95/cloud-pine/compare/v1.0.0...v2.0.0) (2023-05-18)